requirements.txt
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
PyYAML==5.1.2
six==1.13.0
tensorboardX==1.9
torch==1.2.0
torchvision==0.4.0
tqdm==4.39.0
yacs==0.1.6
